final changes

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const Product = require("../models/product");
+const fileHelper = require('../util/file');
 
 const { validationResult } = require('express-validator/check')
 
@@ -7,9 +8,9 @@ exports.getAddProduct = (req, res, next) => {
     pageTitle: "Add Product",
     path: "/admin/add-product",
     editing: false,
-    hasError : false,
-    validationErrors : [],
-    errorMessage : null
+    hasError: false,
+    validationErrors: [],
+    errorMessage: null
   });
 };
 
@@ -18,46 +19,47 @@ exports.postAddProduct = (req, res, next) => {
   const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
-  
-  if(!image){
-    return res.render("admin/edit-product", {
+
+  if (!image) {
+    return res.status(422).render("admin/edit-product", {
       pageTitle: "Add Product",
       path: "/admin/edit-product",
       editing: false,
-      errorMessage : 'Attached file is not a image (allowed types png, jpeg, jpg)',
-      product : {
-        title : title,
-        price : price,
-        description : description
+      errorMessage: 'Attached file is not a image (allowed types png, jpeg, jpg)',
+      product: {
+        title: title,
+        price: price,
+        description: description
       },
-      hasError : true,
-      validationErrors : []
+      hasError: true,
+      validationErrors: []
     });
   }
   const errors = validationResult(req)
-  if(!errors.isEmpty()){
+  if (!errors.isEmpty()) {
     return res.render("admin/edit-product", {
       pageTitle: "Add Product",
       path: "/admin/edit-product",
       editing: false,
-      errorMessage : errors.array()[0].msg,
-      product : {
-        title : title,
-        price : price,
-        description : description
+      errorMessage: errors.array()[0].msg,
+      product: {
+        title: title,
+        price: price,
+        description: description
       },
-      hasError : true,
-      validationErrors : errors.array()
+      hasError: true,
+      validationErrors: errors.array()
     });
   }
 
-  const imageUrl = image.path;
+  const imageUrl = image.path;//path to the file to be stored in db
+
   const product = new Product({
-    title : title,
-    imageUrl : imageUrl,
-    price : price,
-    description : description,
-    userId : req.user // mongoose will pick the _id from the object and assign --> explicitly req.user._id
+    title: title,
+    imageUrl: imageUrl,
+    price: price,
+    description: description,
+    userId: req.user // mongoose will pick the _id from the object and assign --> explicitly req.user._id
   });
   product
     .save()// save method provided by mongoose
@@ -89,9 +91,9 @@ exports.getEditProduct = (req, res, next) => {
         path: "/admin/edit-product",
         editing: editMode,
         product: product,
-        hasError : false,
-        errorMessage : null,
-        validationErrors : []
+        hasError: false,
+        errorMessage: null,
+        validationErrors: []
       });
     })
     .catch(err => {
@@ -109,40 +111,41 @@ exports.postEditProduct = (req, res, next) => {
   const updatedDesc = req.body.description;
 
   const errors = validationResult(req)
-  if(!errors.isEmpty()){
+  if (!errors.isEmpty()) {
     return res.render("admin/edit-product", {
       pageTitle: "Add Product",
       path: "/admin/edit-product",
       editing: true,
-      errorMessage : errors.array()[0].msg,
-      product : {
-        title : updatedTitle,
-        price : updatedPrice,
-        description : updatedDesc,
-        _id : prodId
+      errorMessage: errors.array()[0].msg,
+      product: {
+        title: updatedTitle,
+        price: updatedPrice,
+        description: updatedDesc,
+        _id: prodId
       },
-      hasError : true,
-      validationErrors : errors.array()
+      hasError: true,
+      validationErrors: errors.array()
     });
   }
 
   Product.findById(prodId)
-  .then(product =>{
-    if(product.userId.toString() !== req.user._id.toString()){
-      return res.redirect('/')
-    }
-    product.title = updatedTitle;
-    product.price = updatedPrice;
-    product.description = updatedDesc;
-    if(updatedImage){
-      product.imageUrl = updatedImage.path;
-    }
-    return product.save()
-    .then(result => {
-      console.log("Product Updated");
-      res.redirect("/admin/products");
+    .then(product => {
+      if (product.userId.toString() !== req.user._id.toString()) {
+        return res.redirect('/')
+      }
+      product.title = updatedTitle;
+      product.price = updatedPrice;
+      product.description = updatedDesc;
+      if (updatedImage) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = updatedImage.path;
+      }
+      return product.save()
+        .then(result => {
+          console.log("Product Updated");
+          res.redirect("/admin/products");
+        })
     })
-  })
     .catch(err => {
       const error = new Error(err);
       error.httpStatusCode = 500;
@@ -151,8 +154,8 @@ exports.postEditProduct = (req, res, next) => {
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.find({userId : req.user._id}) // find() -> mongoose method
-  // .select and .populate 
+  Product.find({ userId: req.user._id }) // find() -> mongoose method
+    // .select and .populate 
     .then(products => {
       res.render("admin/products", {
         prods: products,
@@ -167,16 +170,20 @@ exports.getProducts = (req, res, next) => {
     });
 };
 
-exports.postDeleteProduct = (req, res, next) => {
-  const prodId = req.body.productId;
-    Product.deleteOne({userId:req.user._id, _id:prodId}) // findByIdAndremove -> mongoose static method
-    .then(result => {
+exports.deleteProduct = (req, res, next) => {
+  const prodId = req.params.productId;
+  Product.findById(prodId).then(product => {
+    if (!product) {
+      next(new Error('No product found'));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return Product.deleteOne({ userId: req.user._id, _id: prodId }) // findByIdAndremove -> mongoose static method    
+  })
+    .then(() => {
       console.log("Product deleted!!");
-      res.redirect("/admin/products");
+      res.status(200).json({message : 'success!'});
     })
     .catch(err => {
-      const error = new Error(err);
-      error.httpStatusCode = 500;
-      return next(error)
+      res.status(500).json({message : 'deleting product failed'})
     });
 };
